task user reln

+++ todolist_app/views.py
@@ -12,11 +12,17 @@ def todolist(request):
     if request.method == "POST":
         form = TaskForm(request.POST or None)
         if form.is_valid():
-            form.save()
+            instance = form.save(commit=False)
+            instance.manage = request.user
+            instance.save()
+
+    # delay form saving, access manage(user)
+    # save instance
         messages.success(request, ("New Task Added Successfully!"))
         return redirect('todolist')
     else:
-        all_tasks = TaskList.objects.all()
+        all_tasks = TaskList.objects.filter(manage=request.user)
+    #  add filter for task of which user
         paginator = Paginator(all_tasks, 5)
 # paginator instance = paginator class(on which obj pagination, task show on one page)
         page = request.GET.get('pg')
@@ -28,15 +34,23 @@ def todolist(request):
 @login_required
 def delete_task(request, task_id):
     task = TaskList.objects.get(pk=task_id)
-    task.delete()
+    if task.manage == request.user:
+        task.delete()
+    else:
+        messages.error(request, ("Access Restricted. You are not allowed!"))
+
     return redirect('todolist')
 
 
 @login_required
 def complete_task(request, task_id):
     task = TaskList.objects.get(pk=task_id)
-    task.done = True
-    task.save()
+    if task.manage == request.user:
+        task.done = True
+        task.save()
+    else:
+        messages.error(request, ("Access Restricted. You are not allowed!"))
+
     return redirect('todolist')
